Moved shared Renderer.save implementation into baseclass
Also enables saving static html file of widgets
